terraform/scripts/requirements.txt
google-cloud-core==2.3.3
google-cloud-orchestration-airflow==1.9.2
google-cloud-storage==2.12.0